ENT-14 | Add CODEOWNERS file.
Use a CODEOWNERS file to define individuals or teams that are responsible for code in a repository.

Code owners are automatically requested for review when someone opens a pull request that modifies code that they own. When someone with admin permissions has enabled required reviews, they can optionally require approval from a code owner.

To use a CODEOWNERS file, create a new file called CODEOWNERS in the root or .github/ directory of the repository.

See: https://help.github.com/articles/about-codeowners/